+++ vm/runtime/Cargo.toml
@@ -17,3 +17,7 @@ commcid = { path = "../../utils/commcid" }
 filecoin-proofs-api = { git = "https://github.com/filecoin-project/rust-filecoin-proofs-api", rev = "b00d2e26c68e49b81e434739336383304b293395" }
 base64 = "0.12.0"
 neptune = "=0.5.4"
+
+[dev-dependencies]
+interpreter = { path = "../interpreter/"}
+db = { path = "../../node/db/" }
