+++ vm/runtime/src/lib.rs
@@ -251,99 +251,3 @@ pub enum ConsensusFaultType {
     ParentGrinding = 2,
     TimeOffsetMining = 3,
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use cid::{multihash, Codec};
-    use commcid::FilecoinMultihashCode;
-
-    #[derive(Copy, Debug, Clone)]
-    struct TestSyscalls;
-    impl Syscalls for TestSyscalls {
-        fn verify_signature(
-            &self,
-            _signature: &Signature,
-            _signer: &Address,
-            _plaintext: &[u8],
-        ) -> Result<(), Box<dyn StdError>> {
-            Ok(())
-        }
-        fn hash_blake2b(&self, _data: &[u8]) -> Result<[u8; 32], Box<dyn StdError>> {
-            Ok([0u8; 32])
-        }
-        fn compute_unsealed_sector_cid(
-            &self,
-            _reg: RegisteredProof,
-            _pieces: &[PieceInfo],
-        ) -> Result<Cid, Box<dyn StdError>> {
-            Ok(Default::default())
-        }
-        fn verify_seal(&self, vi: &SealVerifyInfo) -> Result<(), Box<dyn StdError>> {
-            let commd = cid_to_data_commitment_v1(&vi.unsealed_cid)?;
-            let commr = cid_to_replica_commitment_v1(&vi.on_chain.sealed_cid)?;
-            let miner_addr = Address::new_id(vi.sector_id.miner);
-            let miner_payload = miner_addr.payload_bytes();
-            let mut prover_id = ProverId::default();
-            prover_id[..miner_payload.len()].copy_from_slice(&miner_payload);
-
-            if !proofs_verify_seal(
-                vi.on_chain.registered_proof.into(),
-                commr,
-                commd,
-                prover_id,
-                SectorId::from(vi.sector_id.number),
-                vi.randomness,
-                vi.interactive_randomness,
-                &vi.on_chain.proof,
-            )? {
-                return Err(format!(
-                    "Invalid proof detected: {:?}",
-                    base64::encode(&vi.on_chain.proof)
-                )
-                .into());
-            }
-
-            Ok(())
-        }
-        fn verify_post(&self, _vi: &WindowPoStVerifyInfo) -> Result<(), Box<dyn StdError>> {
-            Ok(Default::default())
-        }
-        fn verify_consensus_fault(
-            &self,
-            _h1: &[u8],
-            _h2: &[u8],
-            _extra: &[u8],
-            _earliest: ChainEpoch,
-        ) -> Result<Option<ConsensusFault>, Box<dyn StdError>> {
-            Ok(Some(ConsensusFault {
-                target: Address::new_id(0),
-                epoch: 0,
-                fault_type: ConsensusFaultType::DoubleForkMining,
-            }))
-        }
-    }
-
-    #[test]
-    fn verify_seal_test() {
-        let sys = TestSyscalls;
-        let mut vi = SealVerifyInfo::default();
-
-        let data: &[u8; 32] = &[2; 32];
-        let mh_sealed = multihash::wrap(
-            multihash::Code::Custom(FilecoinMultihashCode::SealedV1 as u64),
-            data,
-        );
-        let mh_unsealed = multihash::wrap(
-            multihash::Code::Custom(FilecoinMultihashCode::UnsealedV1 as u64),
-            data,
-        );
-
-        vi.on_chain.sealed_cid = Cid::new_v1(Codec::Raw, mh_sealed);
-        vi.unsealed_cid = Cid::new_v1(Codec::Raw, mh_unsealed);
-        vi.on_chain.registered_proof = RegisteredProof::StackedDRG2KiBSeal;
-
-        // TODO currently captures an error resulting from rust-fil-proofs; need to revisit
-        assert_eq!(sys.verify_seal(&vi).is_err(), true);
-    }
-}
